RawMessageHeader, reading length should be long and not short as before.

+++ src/Vlingo.Wire/Message/RawMessageHeader.cs
@@ -72,7 +72,7 @@ namespace Vlingo.Wire.Message
 
                 var nodeId = binaryReader.ReadInt16();
                 var type = binaryReader.ReadInt16();
-                var length = binaryReader.ReadInt16();
+                var length = binaryReader.ReadInt64();
                 binaryReader.ReadInt16(); // unused1
                 binaryReader.ReadInt16(); // unused2
 
@@ -115,7 +115,7 @@ namespace Vlingo.Wire.Message
             return $"RawMessageHeader[headerId={HeaderId} nodeId={_nodeId} type={_type} length={_length}]";
         }
 
-        protected RawMessageHeader SetAll(short nodeId, short type, int length)
+        protected RawMessageHeader SetAll(short nodeId, short type, long length)
         {
             _nodeId = nodeId;
             _type = type;
